Remove JWT from login/create outputs
